fixup! JAMES-4057 Enable fuzziness search
turn fuzzy search off by default.
