+++ docs/modules/servers/pages/distributed/configure/opensearch.adoc
@@ -98,9 +98,9 @@ If the message is not found, it will fall back to the old behavior (The message 
 Default to false.
 
 | opensearch.text.fuzziness.search
-| Use fuzziness on text searches. This option helps to correct user typing mistakes.
+| Use fuzziness on text searches. This option helps to correct user typing mistakes and makes the result a bit more flexible.
 
-Default to true.
+Default to false.
 
 | opensearch.indexBody
 | Indicates if you wish to index body or not (default: true). This can be used to decrease the performance cost associated with indexing.
