+++ mailbox/opensearch/src/test/java/org/apache/james/mailbox/opensearch/OpenSearchMailboxConfigurationTest.java
@@ -278,10 +278,10 @@ class OpenSearchMailboxConfigurationTest {
     }
 
     @Test
-    void textFuzzinessSearchShouldReturnTrueByDefault() {
+    void textFuzzinessSearchShouldReturnFalseByDefault() {
         PropertiesConfiguration configuration = new PropertiesConfiguration();
 
         assertThat(OpenSearchMailboxConfiguration.fromProperties(configuration).textFuzzinessSearchEnable())
-            .isTrue();
+            .isFalse();
     }
 }
