+++ server/apps/distributed-app/sample-configuration/opensearch.properties
@@ -99,6 +99,6 @@ opensearch.indexAttachments=true
 # When set to true, James will attempt to reindex from the indexed message when moved. If the message is not found, it will fall back to the old behavior (The message will be indexed from the blobStore source)
 # opensearch.message.index.optimize.move=false
 
-# Use fuzziness on text searches. This option helps to correct user typing mistakes. cf: https://opensearch.org/docs/latest/query-dsl/full-text/match/#fuzziness
-# Optional. Defaults to `true`.
-# opensearch.text.fuzziness.search=true
+# Use fuzziness on text searches. This option helps to correct user typing mistakes and makes the result a bit more flexible. cf: https://opensearch.org/docs/latest/query-dsl/full-text/match/#fuzziness
+# Optional. Defaults to `false`.
+# opensearch.text.fuzziness.search=false
